refactor: rename the folder name
